pyproject.toml
[project]
name = "pytorch-unet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib==3.6.2",
    "numpy==1.23.5",
    "pillow==11.3.0",
    "torch==2.9.0",
    "torchaudio==2.9.0",
    "torchvision==0.24.0",
    "tqdm==4.66.3",
    "wandb==0.23.1",
]
